chore: move credential migration logic to deparate class
