+++ packages/auth/amplify_auth_cognito/lib/src/auth_plugin_impl.dart
@@ -15,10 +15,7 @@
 import 'dart:async';
 import 'dart:io';
 
-import 'package:amplify_auth_cognito/src/credentials/legacy_credential_store_data_extension.dart';
-import 'package:amplify_auth_cognito/src/credentials/legacy_ios_cognito_keys.dart';
-import 'package:amplify_auth_cognito/src/credentials/legacy_ios_secure_storage_factory.dart';
-import 'package:amplify_auth_cognito/src/credentials/secure_storage_extension.dart';
+import 'package:amplify_auth_cognito/src/credentials/legacy_credential_provider_impl.dart';
 import 'package:amplify_auth_cognito/src/native_auth_plugin.dart';
 import 'package:amplify_auth_cognito_dart/amplify_auth_cognito_dart.dart';
 import 'package:amplify_auth_cognito_dart/src/flows/hosted_ui/hosted_ui_platform_stub.dart'
@@ -27,7 +24,6 @@ import 'package:amplify_auth_cognito_dart/src/flows/hosted_ui/hosted_ui_platform
 import 'package:amplify_auth_cognito_dart/src/state/machines/hosted_ui_state_machine.dart';
 import 'package:amplify_core/amplify_core.dart';
 import 'package:amplify_secure_storage/amplify_secure_storage.dart';
-import 'package:async/async.dart';
 import 'package:flutter/services.dart';
 
 /// {@template amplify_auth_cognito.amplify_auth_cognito}
@@ -67,6 +63,11 @@ class AmplifyAuthCognito extends AmplifyAuthCognitoDart with AWSDebuggable {
 
     final nativeBridge = NativeAuthBridge();
     stateMachine.addInstance(nativeBridge);
+
+    final legacyCredentialProvider = LegacyCredentialProviderImpl(stateMachine);
+    stateMachine.addInstance<LegacyCredentialProvider>(
+      legacyCredentialProvider,
+    );
     try {
       await nativeBridge.addPlugin();
     } on PlatformException catch (e) {
@@ -114,16 +115,11 @@ class AmplifyAuthCognito extends AmplifyAuthCognitoDart with AWSDebuggable {
 }
 
 class _NativeAmplifyAuthCognito
-    with LegacyIOSSecureStorageProvider, AWSDebuggable, AmplifyLoggerMixin
-    implements NativeAuthPlugin, LegacyCredentialProvider {
-  _NativeAmplifyAuthCognito(this._basePlugin, this._stateMachine) {
-    _stateMachine.addInstance<LegacyCredentialProvider>(this);
-  }
-
+    with AWSDebuggable, AmplifyLoggerMixin
+    implements NativeAuthPlugin {
+  _NativeAmplifyAuthCognito(this._basePlugin, this._stateMachine);
   final AmplifyAuthCognito _basePlugin;
   final CognitoAuthStateMachine _stateMachine;
-
-  final _bundleIdMemoizer = AsyncMemoizer<String>();
 
   @override
   Future<NativeAuthSession> fetchAuthSession(
@@ -174,173 +170,6 @@ class _NativeAmplifyAuthCognito
     } else {
       // Cache them as initial route parameters.
       _stateMachine.addInstance(oauthParameters);
-    }
-  }
-
-  FutureOr<String> _getBundleId() {
-    return _bundleIdMemoizer.runOnce(() {
-      final bridge = _stateMachine.expect<NativeAuthBridge>();
-      return bridge.getBundleId();
-    });
-  }
-
-  @override
-  Future<CredentialStoreData?> fetchLegacyCredentials({
-    CognitoUserPoolConfig? userPoolConfig,
-    CognitoIdentityCredentialsProvider? identityPoolConfig,
-    CognitoOAuthConfig? hostedUiConfig,
-  }) async {
-    if (zIsWeb || !(Platform.isIOS || Platform.isAndroid)) return null;
-    if (Platform.isAndroid) {
-      final bridge = _stateMachine.expect<NativeAuthBridge>();
-      final userPoolConfig = _stateMachine.expect<CognitoUserPoolConfig>();
-      final identityPoolConfig =
-          _stateMachine.expect<CognitoIdentityPoolConfig>();
-      final legacyCredentials = await bridge.getLegacyCredentials(
-        identityPoolConfig.poolId,
-        userPoolConfig.appClientId,
-      );
-      return legacyCredentials.toCredentialStoreData();
-    }
-    final bundleId = await _getBundleId();
-    CognitoUserPoolTokens? userPoolTokens;
-    if (userPoolConfig != null) {
-      final userPoolStorage = getUserPoolStorage(bundleId);
-      final cognitoUserKeys = LegacyCognitoUserKeys(userPoolConfig);
-      final currentUserId = await userPoolStorage.read(
-        key: cognitoUserKeys[LegacyCognitoKey.currentUser],
-      );
-      if (currentUserId != null) {
-        final userPoolKeys = LegacyCognitoUserPoolKeys(
-          currentUserId,
-          userPoolConfig,
-        );
-        final accessToken = await userPoolStorage.read(
-          key: userPoolKeys[LegacyCognitoUserPoolKey.accessToken],
-        );
-        final refreshToken = await userPoolStorage.read(
-          key: userPoolKeys[LegacyCognitoUserPoolKey.refreshToken],
-        );
-        final idToken = await userPoolStorage.read(
-          key: userPoolKeys[LegacyCognitoUserPoolKey.idToken],
-        );
-        if (accessToken != null && refreshToken != null && idToken != null) {
-          // TODO(Jordan-Nelson): fetch sign in method from keychain on iOS
-          final signInMethod = hostedUiConfig != null
-              ? CognitoSignInMethod.hostedUi
-              : CognitoSignInMethod.default$;
-          userPoolTokens = CognitoUserPoolTokens(
-            signInMethod: signInMethod,
-            accessToken: JsonWebToken.parse(accessToken),
-            refreshToken: refreshToken,
-            idToken: JsonWebToken.parse(idToken),
-          );
-        }
-      }
-    }
-
-    String? identityId;
-    AWSCredentials? awsCredentials;
-    final identityPoolId = identityPoolConfig?.poolId;
-    if (identityPoolId != null) {
-      final identityPoolStorage = getIdentityPoolStorage(
-        bundleId,
-        identityPoolId,
-      );
-      const identityPoolKeys = LegacyCognitoIdentityPoolKeys();
-      identityId = await identityPoolStorage.read(
-        key: identityPoolKeys[LegacyCognitoIdentityPoolKey.identityId],
-      );
-      final accessKeyId = await identityPoolStorage.read(
-        key: identityPoolKeys[LegacyCognitoIdentityPoolKey.accessKey],
-      );
-      final secretAccessKey = await identityPoolStorage.read(
-        key: identityPoolKeys[LegacyCognitoIdentityPoolKey.secretKey],
-      );
-      final sessionToken = await identityPoolStorage.read(
-        key: identityPoolKeys[LegacyCognitoIdentityPoolKey.sessionKey],
-      );
-      final expirationStr = await identityPoolStorage.read(
-        key: identityPoolKeys[LegacyCognitoIdentityPoolKey.expiration],
-      );
-      if (accessKeyId != null && secretAccessKey != null) {
-        DateTime? expiration;
-        if (expirationStr != null) {
-          final secondsSinceEpoch = double.tryParse(expirationStr)?.toInt();
-          if (secondsSinceEpoch != null) {
-            expiration = DateTime.fromMillisecondsSinceEpoch(
-              secondsSinceEpoch * 1000,
-              isUtc: true,
-            );
-          }
-        }
-        awsCredentials = AWSCredentials(
-          accessKeyId,
-          secretAccessKey,
-          sessionToken,
-          expiration,
-        );
-      }
-    }
-
-    if ((userPoolTokens ?? awsCredentials ?? identityId) != null) {
-      return CredentialStoreData(
-        userPoolTokens: userPoolTokens,
-        awsCredentials: awsCredentials,
-        identityId: identityId,
-      );
-    }
-    return null;
-  }
-
-  @override
-  Future<void> deleteLegacyCredentials({
-    CognitoUserPoolConfig? userPoolConfig,
-    CognitoIdentityCredentialsProvider? identityPoolConfig,
-    CognitoOAuthConfig? hostedUiConfig,
-  }) async {
-    if (zIsWeb || !(Platform.isIOS || Platform.isAndroid)) return;
-    if (Platform.isAndroid) {
-      final bridge = _stateMachine.expect<NativeAuthBridge>();
-      return bridge.clearLegacyCredentials();
-    }
-    final bundleId = await _getBundleId();
-    if (userPoolConfig != null) {
-      final userPoolStorage = getUserPoolStorage(
-        bundleId,
-      );
-      final cognitoUserKeys = LegacyCognitoUserKeys(userPoolConfig);
-      final currentUser = await userPoolStorage.read(
-        key: cognitoUserKeys[LegacyCognitoKey.currentUser],
-      );
-      if (currentUser != null) {
-        final userPoolKeys = LegacyCognitoUserPoolKeys(
-          currentUser,
-          userPoolConfig,
-        );
-        await userPoolStorage.deleteMany([
-          userPoolKeys[LegacyCognitoUserPoolKey.accessToken],
-          userPoolKeys[LegacyCognitoUserPoolKey.refreshToken],
-          userPoolKeys[LegacyCognitoUserPoolKey.idToken],
-          cognitoUserKeys[LegacyCognitoKey.currentUser],
-        ]);
-      }
-    }
-
-    final identityPoolId = identityPoolConfig?.poolId;
-    if (identityPoolId != null) {
-      final identityPoolStorage = getIdentityPoolStorage(
-        bundleId,
-        identityPoolId,
-      );
-      const identityPoolKeys = LegacyCognitoIdentityPoolKeys();
-      await identityPoolStorage.deleteMany([
-        identityPoolKeys[LegacyCognitoIdentityPoolKey.identityId],
-        identityPoolKeys[LegacyCognitoIdentityPoolKey.accessKey],
-        identityPoolKeys[LegacyCognitoIdentityPoolKey.secretKey],
-        identityPoolKeys[LegacyCognitoIdentityPoolKey.sessionKey],
-        identityPoolKeys[LegacyCognitoIdentityPoolKey.expiration],
-      ]);
     }
   }
